requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
groq==1.0.0
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
python-multipart==0.0.9
psycopg2-binary
